clean up

+++ postGIS_tools/routines/back_up_entire_machine.py
@@ -1,36 +1,27 @@
 """
 TODO: docstrings all across file
 
+
+
+Examples
+--------
+
+    >>> from postGIS_tools.configurations import get_postGIS_config, USER_DESKTOP
+    >>> config, _ = get_postGIS_config()
+    >>> config["localhost"]["debug"] = True
+    >>> back_up_all_databases(USER_DESKTOP, **config["localhost"])
+    >>> remove_all_databases(databases_to_keep=["aaron", "postgres"], **config["localhost"])
+
 """
 
 import psycopg2
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 
-import postGIS_tools
+import postGIS_tools as pGIS
 from postGIS_tools.constants import PG_PASSWORD
 
 
 
-def get_database_list(
-        host: str = 'localhost',
-        username: str = 'postgres',
-        password: str = PG_PASSWORD,
-        port: int = 5432,
-        debug: bool = True
-
-):
-    config = {'host': host, 'username': username, 'password': password, 'port': port, 'debug': debug}
-
-    # Get a list of databases that aren't 'postgres'
-    q = """ SELECT datname FROM pg_database 
-            WHERE datistemplate = false
-                AND datname != 'postgres'; """
-
-    # Run this query from within the 'postgres' db
-    query_result = postGIS_tools.functions.fetch_things_from_database(q, 'postgres', **config)
-    db_list = [x[0] for x in query_result]
-
-    return db_list
 
 
 def back_up_all_databases(
@@ -45,42 +36,43 @@ def back_up_all_databases(
 
     config = {'host': host, 'username': username, 'password': password, 'port': port, 'debug': debug}
 
-    for database in get_database_list(**config):
-        postGIS_tools.functions.dump_database_to_sql_file(database, backup_folder, **config)
+    for database in pGIS.get_database_list(**config):
+        pGIS.dump_database_to_sql_file(database, backup_folder, **config)
 
 
 def remove_all_databases(
+        databases_to_keep: list = ["postgres"],
         host: str = 'localhost',
         username: str = 'postgres',
         password: str = PG_PASSWORD,
         port: int = 5432,
         debug: bool = True
 ):
-    """ DROP each non-postgres database. Proceed with caution """
+    """ DROP each database that's not in the list of databases_to_keep """
     config = {'host': host, 'username': username, 'password': password, 'port': port, 'debug': debug}
 
-    for db_name in get_database_list(**config):
-        query = f"DROP DATABASE {db_name}"
+    for db_name in pGIS.get_database_list(**config):
+        if db_name not in databases_to_keep:
 
-        if debug:
-            print(query)
-        
-        connection = psycopg2.connect(dbname='postgres',
-                                      user='postgres',
-                                      host=host,
-                                      password=password)
+            query = f"DROP DATABASE {db_name}"
 
-        connection.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
+            if debug:
+                print(query)
 
-        cursor = connection.cursor()
-        cursor.execute(query)
-        cursor.close()
-        connection.commit()
-        connection.close()
+            connection = psycopg2.connect(dbname='postgres',
+                                          user='postgres',
+                                          host=host,
+                                          password=password)
+
+            connection.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
+
+            cursor = connection.cursor()
+            cursor.execute(query)
+            cursor.close()
+            connection.commit()
+            connection.close()
+
 
 if __name__ == "__main__":
-    from postGIS_tools.configurations import get_postGIS_config, USER_DESKTOP
-    config, _ = get_postGIS_config()
-    config["localhost"]["debug"] = True
-
-    back_up_all_databases(USER_DESKTOP, **config["localhost"])
+    pass
+    # remove_all_databases(["postgres", "aaron"])
